moved compression to its own package
