+++ go/common/compression/data_compression_service.go
@@ -1,4 +1,4 @@
-package crypto
+package compression
 
 import (
 	"bytes"
